separate CStakeman to own file and re-org dependant files
