+++ src/Makefile.am
@@ -230,6 +230,7 @@ BITCOIN_CORE_H = \
   script/standard.h \
   shutdown.h \
   signet.h \
+  staker.h \
   streams.h \
   support/allocators/secure.h \
   support/allocators/zeroafterfree.h \
@@ -387,6 +388,7 @@ libbitcoin_server_a_SOURCES = \
   script/sigcache.cpp \
   shutdown.cpp \
   signet.cpp \
+  staker.cpp \
   timedata.cpp \
   torcontrol.cpp \
   txdb.cpp \
