+++ src/miner.h
@@ -7,19 +7,30 @@
 #ifndef BITCOIN_MINER_H
 #define BITCOIN_MINER_H
 
+#include <amount.h>
+#include <policy/feerate.h>
 #include <primitives/block.h>
-#include <threadinterrupt.h>
+#include <threadsafety.h>
 #include <txmempool.h>
-#include <validation.h>
-#include <wallet/coincontrol.h>
-#include <wallet/wallet.h>
+#include <util/translation.h>
 
+#include <atomic>
+#include <bits/stdint-uintn.h>
 #include <memory>
 #include <optional>
-#include <stdint.h>
+#include <stddef.h>
+#include <string>
+#include <thread>
+#include <vector>
 
-#include <boost/multi_index_container.hpp>
 #include <boost/multi_index/ordered_index.hpp>
+#include <boost/multi_index_container.hpp>
+
+class CConnman;
+class ChainstateManager;
+namespace interfaces {
+class Chain;
+} /* namespace interfaces */
 
 class CBlockIndex;
 class CChainParams;
@@ -29,6 +40,7 @@ class CWallet;
 
 namespace Consensus { struct Params; };
 
+// logging defaults
 static const bool DEFAULT_PRINTPRIORITY = false;
 
 struct CBlockTemplate
@@ -219,57 +231,5 @@ void InitStakeWallet();
 void SetStakingActive(bool active);
 
 void StakeWallet(interfaces::Chain& chain, const std::string& name, std::optional<bool> load_on_start, std::vector<bilingual_str>& warnings);
-
-class CClientUIInterface;
-
-class CStakeman
-{
-public:
-    struct Options {
-        CClientUIInterface* uiInterface = nullptr;
-        ChainstateManager* chainman = nullptr;
-        CConnman* connman = nullptr;
-        CTxMemPool* mempool = nullptr;
-    };
-
-    CStakeman(bool stake_active = true);
-    ~CStakeman();
-
-    void Init(const Options& connOptions);
-    void InitWallets();
-    bool Start();
-    bool Start(CScheduler& scheduler, const Options& options);
-    void Interrupt();
-    void StopThreads();
-    void Stop()
-    {
-        StopThreads();
-    };
-    bool GetStakingActive() const { return fStakingActive; };
-    void SetStakingActive(bool active);
-    int GetStakingThreadCount() { return threadStakeMinterGroup.size(); };
-    void static ThreadStaker(CWallet* pwallet, ChainstateManager* chainman, CConnman* connman, CTxMemPool* mempool, std::thread::id thread_id, std::atomic<bool> &running);
-    void StakeWalletAdd(const std::string& walletname);
-    void StakeWalletRemove(const std::string& walletname);
-
-    /**
-     * This is signaled when staking activity should cease.
-     */
-    CThreadInterrupt interruptStake;
-
-private:
-    std::atomic<bool> fStakingActive{true};
-
-    std::vector<std::thread> threadStakeMinterGroup GUARDED_BY(cs_threadStakeMinterGroup);
-    mutable RecursiveMutex cs_threadStakeMinterGroup;
-
-    typedef std::unordered_map<std::string, std::thread::id> ThreadMap;
-    ThreadMap tm_;
-
-    CClientUIInterface* clientInterface;
-    ChainstateManager* chainManager;
-    CConnman* connManager;
-    CTxMemPool* memPool;
-};
 
 #endif // BITCOIN_MINER_H
