+++ src/node/context.cpp
@@ -9,7 +9,7 @@
 #include <interfaces/chain.h>
 #include <net.h>
 #include <net_processing.h>
-#include <miner.h>
+#include <staker.h>
 #include <policy/fees.h>
 #include <scheduler.h>
 #include <txmempool.h>
