+++ src/rpc/mining.cpp
@@ -30,6 +30,7 @@
 #include <script/script.h>
 #include <script/signingprovider.h>
 #include <shutdown.h>
+#include <staker.h>
 #include <txmempool.h>
 #include <univalue.h>
 #include <util/fees.h>
